Update ReadMe
Updated the readme, guys if you see this i added contact info at the bottom. Im updating this because we can use the readme in our portfolio and still keep the code secret. if you guys want to fill anything in please do. if not ill get it done little by little